🚚 chore: auth_info 디렉토리를 base와 local로 분리
- 디렉토리를 base와 local로 분리하여 코드 응집도를 높였음
